requirements.txt
asgiref==3.7.2
attrs==20.3.0
blinker==1.4
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.3.0
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
Django==4.2.6
django-filter==23.5
django-resized==1.0.2
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
docopt==0.6.2
gunicorn==22.0.0
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
jellyfish==0.11.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lib50==3.0.7
markdown2==2.4.10
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.0
pexpect==4.8.0
Pillow==10.1.0
pipreqs==0.4.13
ptyprocess==0.7.0
pycairo==1.26.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
pytz==2023.3.post1
PyYAML==5.4.1
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
submit50==3.1.1
systemd-python==234
termcolor==1.1.0
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.6
wadllib==1.3.6
yarg==0.1.9
zipp==1.0.0
